Unify environment provisioning engine and simplify lifecycle state
