+++ apps/server/src/services/projects/project-deletion.ts
@@ -1,7 +1,7 @@
 import {
   requestEnvironmentRemoval,
   sweepProviderEnvironment,
-} from "../environments/provider-orchestration.js";
+} from "../environments/environment-engine.js";
 import { cancelAbandonedProviderCreations } from "../threads/thread-environment-providers.js";
 import { eq, isNotNull } from "drizzle-orm";
 import {
