+++ apps/server/src/services/threads/thread-startup-store.ts
@@ -1,18 +1,89 @@
 import { z } from "zod";
-import { threads, type DbConnection, type DbTransaction } from "@bb/db";
+import {
+  createThreadProvisioningId,
+  threads,
+  type DbConnection,
+  type DbTransaction,
+} from "@bb/db";
 import { and, eq, isNull, or, sql } from "drizzle-orm";
-import { persistedThreadProvisionContextSchema } from "./thread-provisioning-context.js";
-import type {
-  ThreadProvisionContext,
-  ThreadProvisionProviderAsk,
-} from "./thread-provisioning-context.js";
+import {
+  environmentMachineSelectionSchema,
+  jsonValueSchema,
+  promptInputSchema,
+  resolvedThreadExecutionOptionsSchema,
+  clientTurnRequestIdSchema,
+} from "@bb/domain";
+
+const reuseIntentSchema = z.object({
+  type: z.literal("reuse"),
+  environmentId: z.string().min(1),
+});
+
+const providerIntentSchema = z.object({
+  type: z.literal("provider"),
+  environmentProviderId: z.string().min(1),
+  machine: environmentMachineSelectionSchema,
+  inputs: jsonValueSchema.nullable(),
+  selectionResolved: z.boolean().default(true),
+});
+
+export const threadProvisionEnvironmentIntentSchema = z.discriminatedUnion(
+  "type",
+  [reuseIntentSchema, providerIntentSchema],
+);
+
+export const threadForkDescriptorSchema = z.object({
+  sourceProviderThreadId: z.string().min(1),
+  sourceProviderCheckpointId: z.string().min(1).optional(),
+});
+
+export const threadProvisionCommonPayloadSchema = z.object({
+  clientRequestId: clientTurnRequestIdSchema,
+  environmentIntent: threadProvisionEnvironmentIntentSchema,
+  execution: resolvedThreadExecutionOptionsSchema,
+  fork: threadForkDescriptorSchema.nullable().default(null),
+  input: z.array(promptInputSchema),
+  inputGroups: z.array(z.array(promptInputSchema).min(1)).min(1).optional(),
+  titleProvided: z.boolean(),
+  seedWithoutRun: z.boolean().default(false),
+});
+
+export type ThreadForkDescriptor = z.infer<typeof threadForkDescriptorSchema>;
+export type ThreadProvisionEnvironmentIntent = z.infer<
+  typeof threadProvisionEnvironmentIntentSchema
+>;
+type ThreadProvisionOperationPayload = z.infer<
+  typeof threadProvisionCommonPayloadSchema
+>;
+
+interface ThreadProvisioningState {
+  environmentId: string | null;
+  provisionEventSequence: number | null;
+  provisioningId: string;
+  workspaceReadyEventSequence: number | null;
+}
+
+export interface ThreadProvisionContext {
+  request: ThreadProvisionOperationPayload;
+  state: ThreadProvisioningState;
+}
+
+export const persistedThreadProvisionContextSchema = z.object({
+  request: threadProvisionCommonPayloadSchema,
+  state: z.object({
+    environmentId: z.string().nullable(),
+    provisionEventSequence: z.number().nullable(),
+    provisioningId: z.string(),
+    workspaceReadyEventSequence: z.number().nullable(),
+  }),
+});
 
 const providerSchedules = new Map<
   string,
   {
     provisioningId: string;
     environmentProviderId: string;
-    runtime: ThreadProvisionProviderAsk["runtime"];
+    timer: NodeJS.Timeout;
   }
 >();
 
@@ -22,8 +93,6 @@ export function saveThreadProvisionContext(entry: {
   context: ThreadProvisionContext;
   threadId: string;
 }): void {
-  const ask = entry.context.state.providerAsk;
-  const previous = providerSchedules.get(entry.threadId);
   if (
     !persistThreadProvisionContext(
       entry.db,
@@ -31,32 +100,32 @@ export function saveThreadProvisionContext(entry: {
       entry.context,
       entry.replace,
     )
-  ) {
-    if (
-      ask !== null &&
-      ask.runtime !== previous?.runtime &&
-      ask.runtime.nextAskTimer !== null
-    )
-      clearTimeout(ask.runtime.nextAskTimer);
+  )
     return;
-  }
-  if (previous?.runtime !== ask?.runtime)
+  const existing = providerSchedules.get(entry.threadId);
+  if (
+    existing !== undefined &&
+    (existing.provisioningId !== entry.context.state.provisioningId ||
+      entry.context.request.environmentIntent.type === "reuse")
+  )
     clearThreadProvisionSchedule(entry.threadId);
-  if (ask !== null)
-    providerSchedules.set(entry.threadId, {
-      provisioningId: entry.context.state.provisioningId,
-      environmentProviderId: ask.environmentProviderId,
-      runtime: ask.runtime,
-    });
+}
+
+export function setThreadProvisionSchedule(
+  threadId: string,
+  schedule: {
+    provisioningId: string;
+    environmentProviderId: string;
+    timer: NodeJS.Timeout;
+  },
+): void {
+  clearThreadProvisionSchedule(threadId);
+  providerSchedules.set(threadId, schedule);
 }
 
 export function clearThreadProvisionSchedule(threadId: string): void {
   const schedule = providerSchedules.get(threadId);
-  if (schedule !== undefined) {
-    if (schedule.runtime.nextAskTimer !== null)
-      clearTimeout(schedule.runtime.nextAskTimer);
-    schedule.runtime.nextAskTimer = null;
-  }
+  if (schedule !== undefined) clearTimeout(schedule.timer);
   providerSchedules.delete(threadId);
 }
 
@@ -131,21 +200,19 @@ export function readThreadProvisionContext(
     .parse(value);
   if (header.kind !== "provisioning") return null;
   const context = persistedThreadProvisionContextSchema.parse(value);
-  const schedule = providerSchedules.get(threadId);
+  return context;
+}
+
+export function createThreadStartup(
+  request: ThreadProvisionContext["request"],
+): ThreadProvisionContext {
   return {
-    ...context,
+    request: threadProvisionCommonPayloadSchema.parse(request),
     state: {
-      ...context.state,
-      providerAsk:
-        context.state.providerAsk === null
-          ? null
-          : {
-              ...context.state.providerAsk,
-              runtime:
-                schedule?.provisioningId === context.state.provisioningId
-                  ? schedule.runtime
-                  : { nextAskTimer: null, recheckRequested: false },
-            },
+      environmentId: null,
+      provisioningId: createThreadProvisioningId(),
+      provisionEventSequence: null,
+      workspaceReadyEventSequence: null,
     },
   };
 }
