+++ apps/server/test/services/environments/environment-response.test.ts
@@ -28,18 +28,11 @@ function makeRow(overrides: Partial<EnvironmentRow> = {}): EnvironmentRow {
     teardownStatus: null,
     teardownMessage: null,
     resource: null,
-    provisioningThreadId: null,
-    provisioningAttempt: 0,
-    provisioningPhase: null,
-    provisioningFailedAt: null,
-    provisioningFailure: null,
-    provisioningMessage: null,
-    provisioningTransientFailures: 0,
-    provisioningStep: "",
-    provisioningLog: "",
-    provisioningPathRejected: false,
-    provisioningClaimPath: null,
-    provisioningAttached: false,
+    ownerThreadId: null,
+    attempt: 0,
+    statusMessage: null,
+    pendingLog: "",
+    claimPath: null,
     status: "ready",
     createdAt: 1,
     updatedAt: 2,
